perf(stackframe): 宽槽合并 + 无锁帧 + 内存池复用 + 类型截断显式化
- StackFrame 20+ 类型数组合并为 3 个宽槽（int64_t*/double*/void**）
- 私有帧去掉 pthread_rwlock_init（零锁开销），共享帧才初始化
- 线程本地空闲链表（TLS free list）：帧和内部数组复用，零 malloc
- get_*_ptr 统一返回 int64_t*，消除严格别名违规
- inc 函数接收 int64_t* 并内部显式截断（int8/int16/int32/uint8 等）
- opc_name 导出供 VM 调试日志打印枚举名
- 清空 tests/ 旧测试，新建 basic_types_test.lm 双通道验证通过

+++ kit/runtime/include/lm_value.h
@@ -123,20 +123,20 @@ long long range_to_ll(Value v);       // range 参数转 long long
 
 // ===== 类型化自增自减（直接操作原始指针，零转换开销） =====
 /* 有符号整数 */
-void int_inc(int* v);
-void int8_inc(int8_t* v);
-void int16_inc(int16_t* v);
-void int32_inc(int32_t* v);
+void int_inc(int64_t* v);
+void int8_inc(int64_t* v);
+void int16_inc(int64_t* v);
+void int32_inc(int64_t* v);
 void int64_inc(int64_t* v);
 void long_inc(long* v);
 void short_inc(short* v);
 
 /* 无符号整数 */
-void uint_inc(unsigned int* v);
-void uint8_inc(uint8_t* v);
-void uint16_inc(uint16_t* v);
-void uint32_inc(uint32_t* v);
-void uint64_inc(uint64_t* v);
+void uint_inc(int64_t* v);
+void uint8_inc(int64_t* v);
+void uint16_inc(int64_t* v);
+void uint32_inc(int64_t* v);
+void uint64_inc(int64_t* v);
 void ulong_inc(unsigned long* v);
 void ushort_inc(unsigned short* v);
 void byte_inc(unsigned char* v);
